+++ Control/Joint/Concepts/Lens.hs
@@ -3,7 +3,8 @@ module Control.Joint.Concepts.Lens where
 import "comonad" Control.Comonad (extract)
 
 import Control.Joint.Abilities.Adaptable (adapt)
-import Control.Joint.Effects.Store (Store, pos, peek, retrofit)
+import Control.Joint.Effects.State (Stateful, State (State))
+import Control.Joint.Effects.Store (Store (Store), pos, peek, retrofit)
 
 type Lens s t = s -> Store t s
 
@@ -15,3 +16,6 @@ set lens new = peek new . lens
 
 over :: Lens s t -> (t -> t) -> s -> s
 over lens f = extract . retrofit f . lens
+
+zoom :: Stateful bg t => Lens bg ls -> State ls a -> t a
+zoom lens (State f) = adapt . State $ (\(Store (p, g)) -> (\(x,y) -> (g x, y)) . f $ p) . lens