+++ CHANGELOG.md
@@ -61,3 +61,4 @@
 * Define `pos`, `seek`, `peek` and `retrofit` methods for `Store`
 * Define `Lens`, `view`, `set` and `over` in `Store` module
 * Define `Concepts` umbrella module with `Lens` submodule
+* Define `zoom` method in `Concepts.Lens` module